Phase 2: Move nip02FollowLists (6 files) to commons
Move all 6 nip02FollowLists files to commons module:
- DeclaredFollowsPerOutboxRelay
- DeclaredFollowsPerUsingRelay
- FollowListOutboxOrProxyRelays
- FollowListReusedOutboxOrProxyRelays
- FollowsPerOutboxRelay
- Kind3FollowListState

Replace LocalCache with ICacheProvider, Dispatchers.IO with Dispatchers_IO.
Add checkGetOrCreateUser to ICacheProvider interface.
Leave typealiases at original locations.
