+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip02FollowLists/DeclaredFollowsPerUsingRelay.kt
@@ -20,48 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip02FollowLists
 
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.topNavFeeds.UsingRelayUnwrapper
-import com.vitorpamplona.quartz.nip01Core.core.HexKey
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.emitAll
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.transformLatest
-
-/**
- * Computes a list of all declared relays for each user in the follow list.
- */
-class DeclaredFollowsPerUsingRelay(
-    kind3Follows: Kind3FollowListState,
-    val cache: LocalCache,
-    scope: CoroutineScope,
-) {
-    val calculator = UsingRelayUnwrapper()
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    val flow: StateFlow<Map<NormalizedRelayUrl, Set<HexKey>>> =
-        kind3Follows.flow
-            .transformLatest {
-                emitAll(
-                    calculator.toUsersPerRelayFlow(it.authors, cache) { it },
-                )
-            }.onStart {
-                emit(
-                    calculator.usersPerRelaySnapshot(kind3Follows.flow.value.authors, cache) { it },
-                )
-            }.distinctUntilChanged()
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptyMap(),
-            )
-}
+typealias DeclaredFollowsPerUsingRelay = com.vitorpamplona.amethyst.commons.model.nip02FollowLists.DeclaredFollowsPerUsingRelay
