+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip02FollowLists/FollowListReusedOutboxOrProxyRelays.kt
@@ -20,116 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip02FollowLists
 
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.nip51Lists.blockedRelays.BlockedRelayListState
-import com.vitorpamplona.amethyst.model.nip51Lists.proxyRelays.ProxyRelayListState
-import com.vitorpamplona.amethyst.model.topNavFeeds.OutboxRelayLoader
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.emitAll
-import kotlinx.coroutines.flow.flatMapLatest
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.transformLatest
-
-class FollowListReusedOutboxOrProxyRelays(
-    kind3Follows: Kind3FollowListState,
-    blockedRelayList: BlockedRelayListState,
-    proxyRelayList: ProxyRelayListState,
-    val cache: LocalCache,
-    scope: CoroutineScope,
-) {
-    @OptIn(ExperimentalCoroutinesApi::class)
-    val outboxRelayFlow: StateFlow<Set<NormalizedRelayUrl>> =
-        kind3Follows.flow
-            .transformLatest { follows ->
-                emitAll(
-                    OutboxRelayLoader(true).toAuthorsPerRelayFlow(follows.authors, cache) { authorsPerRelay ->
-                        authorsPerRelay
-                            .mapNotNull {
-                                if (it.value.size > 1) {
-                                    it.key
-                                } else {
-                                    null
-                                }
-                            }.toSet()
-                    },
-                )
-            }.onStart {
-                emit(
-                    OutboxRelayLoader(true).authorsPerRelaySnapshot(kind3Follows.flow.value.authors, cache) { authorsPerRelay ->
-                        authorsPerRelay
-                            .mapNotNull {
-                                if (it.value.size > 1) {
-                                    it.key
-                                } else {
-                                    null
-                                }
-                            }.toSet()
-                    },
-                )
-            }.distinctUntilChanged()
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    val outboxRelayMinusBlockedFlow: StateFlow<Set<NormalizedRelayUrl>> =
-        combine(outboxRelayFlow, blockedRelayList.flow) { followList, blockedRelays ->
-            followList.minus(blockedRelays)
-        }.onStart {
-            emit(outboxRelayFlow.value.minus(blockedRelayList.flow.value.toSet()))
-        }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    val flow: StateFlow<Set<NormalizedRelayUrl>> =
-        proxyRelayList.flow
-            .flatMapLatest { proxyRelays ->
-                if (proxyRelays.isEmpty()) {
-                    outboxRelayMinusBlockedFlow
-                } else {
-                    MutableStateFlow(proxyRelays)
-                }
-            }.onStart {
-                emit(
-                    proxyRelayList.flow.value.ifEmpty {
-                        outboxRelayMinusBlockedFlow.value
-                    },
-                )
-            }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    val flowSet: StateFlow<Set<String>> =
-        flow
-            .map { relayList ->
-                relayList.mapTo(mutableSetOf()) { it.url }
-            }.onStart {
-                emit(flow.value.mapTo(mutableSetOf()) { it.url })
-            }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-}
+typealias FollowListReusedOutboxOrProxyRelays = com.vitorpamplona.amethyst.commons.model.nip02FollowLists.FollowListReusedOutboxOrProxyRelays
