+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip02FollowLists/FollowsPerOutboxRelay.kt
@@ -20,86 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip02FollowLists
 
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.nip51Lists.blockedRelays.BlockedRelayListState
-import com.vitorpamplona.amethyst.model.nip51Lists.proxyRelays.ProxyRelayListState
-import com.vitorpamplona.amethyst.model.topNavFeeds.OutboxRelayLoader
-import com.vitorpamplona.quartz.nip01Core.core.HexKey
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.emitAll
-import kotlinx.coroutines.flow.flatMapLatest
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.transformLatest
-
-class FollowsPerOutboxRelay(
-    kind3Follows: Kind3FollowListState,
-    blockedRelayList: BlockedRelayListState,
-    proxyRelayList: ProxyRelayListState,
-    val cache: LocalCache,
-    scope: CoroutineScope,
-) {
-    @OptIn(ExperimentalCoroutinesApi::class)
-    val outboxPerRelayFlow: StateFlow<Map<NormalizedRelayUrl, Set<HexKey>>> =
-        kind3Follows.flow
-            .transformLatest {
-                emitAll(
-                    OutboxRelayLoader().toAuthorsPerRelayFlow(it.authors, cache) { it },
-                )
-            }.onStart {
-                emit(
-                    OutboxRelayLoader().authorsPerRelaySnapshot(kind3Follows.flow.value.authors, cache) { it },
-                )
-            }.distinctUntilChanged()
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptyMap(),
-            )
-
-    val outboxPerRelayMinusBlockedFlow: StateFlow<Map<NormalizedRelayUrl, Set<HexKey>>> =
-        combine(outboxPerRelayFlow, blockedRelayList.flow) { followList, blockedRelays ->
-            followList.minus(blockedRelays)
-        }.onStart {
-            emit(outboxPerRelayFlow.value.minus(blockedRelayList.flow.value.toSet()))
-        }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptyMap(),
-            )
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    val flow: StateFlow<Map<NormalizedRelayUrl, Set<HexKey>>> =
-        proxyRelayList.flow
-            .flatMapLatest { proxyRelays ->
-                if (proxyRelays.isEmpty()) {
-                    outboxPerRelayMinusBlockedFlow
-                } else {
-                    kind3Follows.flow.map { follows ->
-                        proxyRelays.associateWith { follows.authors }
-                    }
-                }
-            }.onStart {
-                if (proxyRelayList.flow.value.isEmpty()) {
-                    emit(outboxPerRelayMinusBlockedFlow.value)
-                } else {
-                    emit(proxyRelayList.flow.value.associateWith { kind3Follows.flow.value.authors })
-                }
-            }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptyMap(),
-            )
-}
+typealias FollowsPerOutboxRelay = com.vitorpamplona.amethyst.commons.model.nip02FollowLists.FollowsPerOutboxRelay
